Adding a very rough example of using valarray
as the basic data container in Field3D. Only applied in F3D_OP_REAL
currently. To get the most from this need to look at enabling operations
with Field?D etc. to "resolve" to operations on the underlying valarray
container, possibly using expression templates.

Currently expect this code to be slower than using the usual Array
container as it has no reference counting etc. to save on allocations.