Adding initial replacement for Array
that use `std::valarray` as the data container and `std::shared_pointer`
for providing reference counting and similar.

Should act as a drop in replace for `Array` for any acceptable numeric type.

Moderately tested so far, but generally seems to work well. Possible
that there's some code we can remove given that this uses shared
pointers and not raw pointers (a lot of the copying to old etc.)