Adding example `get(Region)` routine for `Field3D`
Requires `c++14` or newer in order to allow `auto` for function return.

This allows us to pass a region (gslice) to `get` in order to return a
subset of the data. Due to the resulting `gslice_array` not having
direct arithmetic operations we must cast to a `valarray` when using on
the RHS, this will copy the data. We avoid this cast when used on the
LHS which just returns a view to the subset.

Generally more efficient to do

`result.get()[reg] = a.get()[reg] + b.get()[reg]`

than

`result.get(reg) = a.get(reg) + b.get(reg)`

because of the copying/casting required on the right in the second version.